Keep DroidCall ingestion source-only
